+++ src/core/src/UI/index.js
@@ -1,11 +1,11 @@
 import { disableRect, enableRect } from './rect';
 import { createSpan, getRectangles } from './span';
 import { createRelation } from './relation';
-import { disableViewMode, enableViewMode } from './view';
+import { enableViewMode } from './view';
 
 export default {
   disableRect, enableRect,
   createSpan, getRectangles,
   createRelation,
-  disableViewMode, enableViewMode
+  enableViewMode
 };
