+++ src/page/ui/annotationTools.js
@@ -7,9 +7,6 @@
 */
 export function setup() {
 
-    // TODO Restoreようだけど必要？
-    window.currentAnnoToolType = 'view';
-
     // Rect annotation button.
     $('.js-tool-btn-rect').off('click').on('click', (e) => {
