+++ src/page/ui/downloadButton.js
@@ -25,7 +25,6 @@ export function setup() {
  */
 function downloadAnnotation() {
 
-    // TODO UIとの分離.
     window.annoPage.exportData().then(annotations => {
         let blob = new Blob([annotations]);
         let blobURL = window.URL.createObjectURL(blob);
