+++ src/page/ui/referenceAnnoDropdown.js
@@ -1,7 +1,6 @@
 /**
  * UI parts - Reference Annotation Dropdown.
  */
-import { displayAnnotation } from '../util/display';
 
 /**
  * Setup a click action of the Reference Annotation Dropdown.
@@ -27,8 +26,7 @@ export function setup() {
             $('#dropdownAnnoReference .js-text').text('Reference Files');
         }
 
-        // TODO Move to PDFAnnoPage.js.
-        displayAnnotation(false);
+        window.annoPage.displayAnnotation(false);
 
         return false;
